+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-30T08:15:15.305Z",
+    "2026-07-01T06:53:32.649Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {